factor out EqualRange matcher to own header
