+++ tests/include/CatchUtils.hpp
@@ -1,0 +1,32 @@
+#include <catch2/catch_test_macros.hpp>
+#include <catch2/matchers/catch_matchers_templated.hpp>
+
+namespace fluid { 
+
+template<typename Range>
+struct EqualsRangeMatcher : Catch::Matchers::MatcherGenericBase {
+    EqualsRangeMatcher(Range&& range):
+        range{ range }
+    {}
+
+    template<typename OtherRange>
+    bool match(OtherRange const& other) const {
+        using std::begin; using std::end;
+
+        return std::equal(begin(range), end(range), begin(other), end(other));
+    }
+
+    std::string describe() const override {
+        return "Equals: " + Catch::rangeToString(range);
+    }
+
+private:
+    Range range;
+};
+
+template<typename Range>
+auto EqualsRange(Range&& range) -> EqualsRangeMatcher<Range> {
+    return EqualsRangeMatcher<Range>{std::forward<Range>(range)};
+}
+
+}
